requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.9.3
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.12.3
billiard==4.2.4
blinker==1.9.0
cachelib==0.13.0
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Flask==2.3.3
Flask-Caching==2.3.1
Flask-Cors==4.0.0
frozenlist==1.8.0
greenlet==3.0.3
gunicorn==21.2.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.6.2
lxml==5.1.0
MarkupSafe==3.0.3
multidict==6.7.1
packaging==26.0
playwright==1.42.0
prompt_toolkit==3.0.52
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
pyee==11.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
pytest==8.3.5
pytest-cov==6.0.0
ratelimit==2.2.1
redis==7.1.1
requests==2.31.0
six==1.17.0
soupsieve==2.8.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
Werkzeug==3.1.5
yarl==1.22.0
